keccak: rename `aarch64_sha3` module to `armv8`
This is more consistent with the other crates we have which implement
ARMv8 intrinsics, such as the `aes` and `polyal` crates.
